+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Support generating accessors for generic types ([#10](https://github.com/nesk/akkurate/issues/10))
+- A KDoc is provided for each validatable accessor ([#6](https://github.com/nesk/akkurate/issues/6))
 
 ## [0.2.0] - 2023-09-26